Separated interface definition from implementation
